Refactor; Also, auto generate filters when code changes.
